change: Refactor MixingConfig to forward Columns
Since the padding of the `next_inputs` is given in base-2,
`AbsorbConfig` has been refactored to get the `next_inputs` as
`AssignedCells` as well as performing the base conversion `2->9` inside.

This means that we need to update how MixingConfig generates
AbsorbConfig forwarding the BaseConversionConfig and the state.
Also, it does the same itself recieving them.

This is the last step needed to integrate the RLC into this config.