feat: add precompile skeleton
This commit introduces the support for precompiled contracts. For more
information, check the appendix E of the Ethereum Yellow Paper.

It is a simplified version of the identity precompile PR of PSE. It
strips the implementation of the identity precompile variant, leaving
only the skeleton for the lookup table calls.

[Ethereum Yellow Paper](https://ethereum.github.io/yellowpaper/paper.pdf)
[Identity (dataCopy) precompile #1396](https://github.com/privacy-scaling-explorations/zkevm-circuits/pull/1396)