Updated BLAKE2b compression function circuit
Cargo.toml:
1. Default comment was removed.
2. "hex = "0.4"" was added instead of "hex = "0.4.3"".

Cargo.lock: Recreated.

blake2.rs: The new file containing the "blake2b" module, almost all content of which (except for the "test" and "benchmark" submodules) is the same as the one of "lib.rs" just before it undergone the last change from the list below. The content of the "test" submodule is related to the one of "integration.rs" in the similar way. The same is true about the relation between the "test" submodule and "bench.rs".

lib.rs:
1. "Chunks" instead of "chunks" at the end of the comments for "fn create_chuncks".
2. "pub" was removed from many places: now the data types of the circuit's basic building blocks are not exported.
3. Trailing spaces were removed.
4. "Thus, 56 * 8 + 32 cells or 240 cell pairs are required" was added after "56 byte-column chunks of height 8 and 32 byte-column cells".
5. The comment "14 bit columns are required: 10 collumns for..." was added.
6. The end of "fn synthesize" was rewritten in order to check the correctness of the computation only if tests are run.
7  The "blake2" submodule was renamed into "compression".
8. The content is moved to the "blake2b" module and replaced with "mod blake2b;".

integration.rs:
1. Trailing spaces were removed.
2. The content is moved to the "test" submodule of the "blake2b" module; the file was deleted.

bench.rs:
1. Trailing spaces were removed.
2. The comment describing the correct way of running "fn bench()" was added.
3. The content is moved to the "benchmark" submodule of the "blake2b" module; the file was deleted.